requirements.txt
openai
pandas
lizard
pyyaml
python-dotenv
jinja2
pyyaml
requests
